% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CI.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculate Confidence Interval}
\usage{
calculate_CI(x, conf)
}
\arguments{
\item{x}{vector, a vector containing numeric values, the confidence interval of which mean being calcualted.}

\item{conf}{floating number, a number within the range of \link{0, 1} ($=1-\alpha$), with $\alpha$ as statistical
significance level.}
}
\value{
a matrix, where the first value of the matrix is the \code{lower_bound}, the second value of the matrix
is the \code{upper_bound}.
}
\description{
A function that computes a confidence interval (CI) (e.g. a 95\% CI) for the estimate of the mean in the
population.
}
\details{
This function computes the confidence interval to the mean value of samples contained in a vector
and the confidence level set by the user.
}
\examples{
x <- c(rnorm(100))
calculate_CI(x, conf = 0.95)

}
