% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_p3_class.R
\name{make_p3_class}
\alias{make_p3_class}
\title{Create \code{p3_class}}
\usage{
make_p3_class(x)
}
\arguments{
\item{x}{vector, a vector to be transformed to \code{p3_class}.}
}
\value{
a \code{p3_class} object.
}
\description{
create a new S3 class called \code{p3_class} (Project 3 class).
}
\details{
This function create a new S3 class called \code{p3_class} to the vector input by the user.
}
\examples{
set.seed(1234)
x <- rnorm(100)
p3 <- make_p3_class(x)
print(p3)         # explicitly using the print() method
## a p3_class with 100 observations
p3                # using autoprinting
## a p3_class with 100 observations

}
